Made kvrange fetching > block_device layer

+++ kernel/include/services/storage/partition.h
@@ -9,20 +9,20 @@ typedef struct stor_device stor_device_t;
 typedef struct __attribute__((packed)) stor_partition_t 
 {
     stor_device_t* device;
-    u64 index; 
+    usize index; 
 
-    u64 block_lba;
-    u64 block_amount;
+    usize block_lba;
+    usize block_amount;
 
 } stor_partition_t;
 
 typedef struct partition_table
 {
     stor_partition_t** arr;
-    u64 count;
+    usize count;
 } partition_table_t;
 
-stor_partition_t* get_partition(stor_device_t* device, u64 index);
+stor_partition_t* get_partition(stor_device_t* device, usize index);
 void device_scan_paritions(stor_device_t* device);
 
 #endif // __PARTITION_H__
